consumer/requirements.txt
websockets>=12.0
anthropic>=0.40.0
rich>=13.0
paho-mqtt>=2.0
